requirements.txt
fastapi[standard]
pydantic
python-multipart
python-dotenv
